requirements.txt
appnope==0.1.4
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
google-api-core==1.33.2
google-auth==2.30.0
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
